requirements.txt
fastapi[all]
uvicorn
Pillow
transformers
torch
